refactor dashboard component
