+++ src/components/dashboard/DeleteConfirmationDialog.tsx
@@ -1,0 +1,55 @@
+'use client';
+
+import { Filament } from '@/types/Filament';
+import {
+  Dialog,
+  DialogTitle,
+  DialogContent,
+  DialogActions,
+  DialogContentText,
+  Button,
+  Box,
+  CircularProgress,
+} from '@mui/material';
+
+interface DeleteConfirmationDialogProps {
+  open: boolean;
+  filament: Filament | null;
+  isLoading: boolean;
+  onClose: () => void;
+  onConfirm: () => void;
+}
+
+export default function DeleteConfirmationDialog({
+  open,
+  filament,
+  isLoading,
+  onClose,
+  onConfirm,
+}: DeleteConfirmationDialogProps) {
+  return (
+    <Dialog open={open} onClose={onClose}>
+      <DialogTitle>Delete Filament</DialogTitle>
+      <DialogContent>
+        <DialogContentText>
+          {`Are you sure you want to delete "${filament?.name}"? This action cannot be undone.`}
+        </DialogContentText>
+      </DialogContent>
+      <DialogActions>
+        <Button onClick={onClose} disabled={isLoading}>
+          Cancel
+        </Button>
+        <Button onClick={onConfirm} color="error" variant="contained" disabled={isLoading}>
+          {isLoading ? (
+            <Box sx={{ display: 'flex', alignItems: 'center' }}>
+              <CircularProgress size={20} color="inherit" sx={{ mr: 1 }} />
+              Deleting...
+            </Box>
+          ) : (
+            'Delete'
+          )}
+        </Button>
+      </DialogActions>
+    </Dialog>
+  );
+}
