+++ src/components/dashboard/LoadingSpinner.tsx
@@ -1,0 +1,18 @@
+'use client';
+
+import { Box, CircularProgress, Typography } from '@mui/material';
+
+interface LoadingSpinnerProps {
+  message?: string;
+}
+
+export default function LoadingSpinner({ message = 'Loading...' }: LoadingSpinnerProps) {
+  return (
+    <Box display="flex" justifyContent="center" alignItems="center" minHeight="100vh">
+      <CircularProgress />
+      <Typography variant="h6" sx={{ ml: 2 }}>
+        {message}
+      </Typography>
+    </Box>
+  );
+}
